Only admin can do admin help.

+++ Bot.py
@@ -206,7 +206,7 @@ class Bot( SingleServerIRCBot ):
 		
 		if cmd == 'help':
 			self.privmsg( target, '!help: this help text' )
-		elif cmd == 'admin_help':
+		elif admin and cmd == 'admin_help':
 			self.notice( source, '!die: kill the bot' )
 			self.notice( source, '!raw: send raw irc command' )
